pyproject.toml
[project]
name = "Main-pj-Ai-Service"
version = "0.1.0"
description = ""
authors = [
    {name = "hak1319",email = "gkr054@naver.com"}
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv>=1.0.1,<2.0.0",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "cryptography (>=44.0.2,<45.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ipython = "^9.0.2"
django-extensions = "^3.2.3"

[tool.isort]
profile = "black"
skip_glob = ["*/migrations/*.py"]

[tool.black]
extend-exclude = '''
/(
    migrations/.*
)/
'''

[tool.poetry]
package-mode = false